fix: check hash of mentor file
When downloading the mentor I accidentally closed my terminal which resulted in an incomplete mentor file. Running the command again just produced an empty string.

This change does a check when 'download_mentor' is called to determine if the local file matches the remote file, and if there is a mismatch the mentor is redownloaded.

The remote hash is hardcoded.